+++ app/Containers/Debugger/ValueObjects/RequestsLogger.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Containers\Debugger\Objects;
+namespace App\Containers\Debugger\ValueObjects;
 
+use App\Ship\Parents\ValueObjects\ValueObject;
 use DB;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Config;
@@ -15,7 +16,7 @@ use Monolog\Logger;
  *
  * @author  Mahmoud Zalt  <mahmoud@zalt.me>
  */
-class RequestsLogger
+class RequestsLogger extends ValueObject
 {
 
     CONST TESTING_ENV = 'testing';
@@ -38,7 +39,7 @@ class RequestsLogger
     }
 
     /**
-     * @param \App\Containers\Debugger\Objects\Output $output
+     * @param \App\Containers\Debugger\ValueObjects\Output $output
      */
     public function releaseOutput(Output $output)
     {