+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 ## [Unreleased]
 
 ### Added
-- ...
+- Added ValueObjects class type to be extended by classes that do not requires to be stored in the DB or have ID.
 
 ### Changed
 - Changed the `Content-Language` header field (for requesting resources in a specific language) to `Accept-Language` instead (cf. [Specs](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language)).
 - Rename `GiveAllPermissionsToRole` to `GiveAllPermissionsToRoleCommand`.
 - The structure of the `supported_languages` in `App/Containers/Localization/Configs` was changed in order to support `regions`.
 - The route `/logout` now uses `DELETE` instead of `POST` (to be more RESTful)
+- Move `Localization` and `Region` from Model to ValueObjects folder in the localization container.
+- Move `Output` and `RequestLogger` from Model to ValueObjects folder in the debugger container.
 
 ### Fixed
 - ... 
@@ -17,6 +19,8 @@
 ### Removed
 - ...
 
+
+___
 
 ## v7.1.0 (2017-10-17)